Move responsibility of merging attributes to the module resolve state
This should make merging faster because we don't need to create an iterator each
time the method is called, like it used to be.
